// src/app/types/androidEnum.ts
export enum ModuleStatus {
  DRAFT = 'draft',
  LIVE = 'live',
  READY = 'ready',
  PUBLISH_PROGRESS = 'publish in progress',
  COMPLETE = 'complete',
  ENDED = 'ended',
  ARCHIVED = 'archived',
}

export enum ModuleValueStatus {
  DFT = 'default',
  SAVED = 'saved',
  PUBLISHED = 'published',
  INCOMPLETE = 'incomplete',
  ARCHIVED = 'archived',
  ENDED = 'ended',
}

export enum AndroidPlatform {
  mobile = 'Mobile',
  tv = '10 Foot',
}